Warning correction
There were comparisons between signed and unsigned integer expressions,
in for loops. Now the loop's iterators are size_t.

+++ String.cpp
@@ -203,10 +203,10 @@ String String::operator=(const char* c){ // c is const because we don't change I
 String operator+(const String& lhs, const String& rhs) {
 
   char* newData=new char[lhs.capacity_+rhs.capacity_ +1];
-  for(int i=0;i<lhs.size_;++i){
+  for(size_t i=0;i<lhs.size_;++i){
     newData[i]=lhs.data_[i];
   }
-  for(int j=0;j<=(rhs.size_);++j){
+  for(size_t j=0;j<=(rhs.size_);++j){
     newData[j+lhs.size_]=rhs.data_[j];
   }
   String concatenate (newData);
